==> Backend/app/preprocessing/2. handle_missing_values-1.ipynb <==
import pandas as pd

df = pd.read_csv(r'../data/cleaned_dataset_step_1.csv')

df.head()

missing_values = df.isnull().sum()
print("Missing values per column: ")
print(missing_values)

# Filling missing values on Hireable
df['Hireable'] = df['Hireable'].fillna(False)

missing_values = df.isnull().sum()
print("Missing values per column after handling 'Hireable': ")
print(missing_values)

# checking if the missing values in Repositories Count, Total Repository Size (KB), Total Stars and Followers are the same
missing_rows = df[df['Total Repository Size (KB)'].isnull() & df['Repositories Count'].isnull() & df['Total Stars'].isnull() & df['Followers'].isnull()]
print("Total count of missing Top Repo, Repo count, total stars and followers: ",len(missing_rows))

# Dropping rows which have all 4 attributes missing
df = df.dropna(subset=['Total Repository Size (KB)','Repositories Count','Total Stars', 'Followers'], how = 'all')

missing_values = df.isnull().sum()
print("Missing values per column after hadling 'Total Repository Size (KB)','Repositories Count','Total Stars' and 'Followers': ")
print(missing_values)

# Dropping rows which have 'Account Created At' row missing
df = df.dropna(subset=['Account Created At'])

missing_values = df.isnull().sum()
print("Missing values per column after hadling 'Account Created At': ")
print(missing_values)

# Manually convert columns to numeric, coercing errors to NaN
df['Repositories Count'] = pd.to_numeric(df['Repositories Count'], errors='coerce')
df['Followers'] = pd.to_numeric(df['Followers'], errors='coerce')
df['Total Stars'] = pd.to_numeric(df['Total Stars'], errors='coerce')
df['Total Repository Size (KB)'] = pd.to_numeric(df['Total Repository Size (KB)'], errors='coerce')

# Convert 'Account Created At' to datetime, coercing errors to NaT (Not a Time)
df['Account Created At'] = pd.to_datetime(df['Account Created At'], errors='coerce')

# Check missing values after conversion
missing_values = df.isnull().sum()
print("Missing values per column after handling conversions:")
print(missing_values)



# checking if the missing values in Repositories Count, Total Repository Size (KB), Total Stars and Followers are the same
missing_rows = df[df['Total Repository Size (KB)'].isnull() & df['Repositories Count'].isnull() & df['Total Stars'].isnull() & df['Followers'].isnull() & df['Account Created At'].isnull()]
print("Total count of missing Top Repo, Repo count, total stars and followers: ",len(missing_rows))

# Dropping rows which have all 4 attributes missing
df = df.dropna(subset=['Total Repository Size (KB)','Repositories Count','Total Stars', 'Followers','Account Created At'], how = 'all')

missing_values = df.isnull().sum()
print("Missing values per column after hadling all coreced values: ")
print(missing_values)

df.to_csv('../data/cleaned_dataset_step_2.csv', index=False)
