==> Backend/app/preprocessing/4. city_mapping_normalization.ipynb <==
import pandas as pd


df = pd.read_csv('../data/cleaned_dataset_step_3.csv')

df.isnull().sum()

city_country_mapping = pd.read_csv('../data/worldcities.csv')

df['city'] = df['Country'].apply(lambda x: x.split(',')[0].strip() if isinstance(x, str) else None)


print(df['city'].unique())

print("Unique cities in df:", df['city'].unique())
print("Unique cities in city_country_mapping:", city_country_mapping['city'].unique())

# Strip spaces and convert to lowercase
df['city'] = df['city'].str.strip().str.lower()
city_country_mapping['city'] = city_country_mapping['city'].str.strip().str.lower()
print(df['city'])
print(city_country_mapping['city'])


city_country_mapping = city_country_mapping.drop_duplicates(subset=['city'], keep='first')

df = df.merge(city_country_mapping, how='left', on='city')
print("Merge result (head):")
print(df.head())


unmatched_cities = df[~df['city'].isin(city_country_mapping['city'])]
print("Unmatched cities:")
print(unmatched_cities['city'].unique())
print(len(unmatched_cities['city'].unique()))

df

df = df.drop(columns=['city','city_ascii',	'lat',	'lng',	'iso2',	'iso3'	,'admin_name',	'capital',	'population',	'id'])


df

df["Country"] = df["country"].combine_first(df["Country"])

df.drop(columns=['country'],inplace=True)

df

df.to_csv('../data/cleaned_dataset_step_4.csv', index=False)
print("Results saved to '('../data/cleaned_dataset_step_4.csv'.")
